+++ examples/resources/pcd_host_cluster_role/import.sh
@@ -1,0 +1,1 @@
+terraform import pcd_host_cluster_role.hypervisor <host_id>/hypervisor